requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
httpx
Pillow
